更新代码，添加文字和箭头功能

+++ CLRCutMosaicDraw/CLRDrawElementModelStore.m
@@ -87,7 +87,7 @@
 
 - (DrawElementsModel *) getModelByNum:(NSInteger)num
 {
-    NSLog(@"%d",num);
+    NSLog(@"%ld",(long)num);
     return [self.modelStoreArray objectAtIndex:num];
 }
